Waitpid in the loop loses data
Weirdly doing waitpid(NOHANG) causes data loss

IN any case the data comes out in a slightly different order with STDERR
mixed in with STDOUT instead of at the end(?)
Which must be some buffering thing.
Not worried.

Also loses all the original coloring, if any. But the other bash-based tool
did as well, so also not worried.